# api/swagger.yaml
basePath: /
definitions:
...
  presenters.DayToDayExpense:
    properties:
      amount:
        type: number
      day:
        type: string
      day_name:
        type: string
      month:
        type: string
    type: object
  presenters.GetAvailableMonthsYearsOutputDto:
    properties:
      available_months:
        items:
          $ref: '#/definitions/repositories.MonthOption'
        type: array
      available_years:
        items:
          type: integer
        type: array
    type: object
  presenters.GetCategoryTagsTotalsByMonthYearOutputDto:
    properties:
      expenses:
        $ref: '#/definitions/repositories.CategoryTagsTotals'
    type: object
  presenters.GetDayToDayExpensesPeriodOutputDto:
    properties:
      expenses:
        items:
          $ref: '#/definitions/presenters.DayToDayExpense'
        type: array
    type: object
  presenters.GetExpensesByCategoryPeriodOutputDto:
    properties:
      expenses:
        items:
          $ref: '#/definitions/repositories.CategoryExpense'
        type: array
    type: object
  presenters.GetExpensesByMonthYearOutputDto:
    properties:
      expenses:
        $ref: '#/definitions/repositories.MonthExpenses'
    type: object
  presenters.GetMonthlyExpensesByCategoryYearOutputDto:
    properties:
      available_years:
        items:
          type: integer
        type: array
      expenses:
        items:
          $ref: '#/definitions/repositories.MonthlyCategoryExpense'
        type: array
    type: object
  presenters.GetMonthlyExpensesByTagYearOutputDto:
    properties:
      available_years:
        items:
          type: integer
        type: array
      expenses:
        items:
          $ref: '#/definitions/repositories.MonthlyTagExpense'
        type: array
    type: object
  presenters.GetTotalExpensesForCurrentMonthOutputDto:
    properties:
      current_month:
        type: string
      total_expenses:
        type: number
    type: object
  presenters.GetTotalExpensesForCurrentWeekOutputDto:
    properties:
      current_week:
        type: string
      total_expenses:
        type: number
    type: object
  presenters.GetTotalExpensesForPeriodOutputDto:
    properties:
      total:
        type: number
    type: object
  presenters.GetTotalExpensesMonthCurrentYearOutputDto:
    properties:
      expenses_month_current_year:
        $ref: '#/definitions/repositories.ExpensesMonthCurrentYear'
    type: object
  repositories.CategoryExpense:
    properties:
      category_color:
        type: string
      category_name:
        type: string
      total:
        type: number
    type: object
  repositories.CategoryTagTotal:
    properties:
      name:
        type: string
      tag_amount:
        type: number
    type: object
  repositories.CategoryTagsTotals:
    properties:
      available_months:
        items:
          $ref: '#/definitions/repositories.MonthOption'
        type: array
      available_years:
        items:
          type: integer
        type: array
      categories:
        items:
          $ref: '#/definitions/repositories.CategoryWithTags'
        type: array
      expenses_amount:
        type: number
      month:
        type: string
      year:
        type: integer
    type: object
  repositories.CategoryWithTags:
    properties:
      category_amount:
        type: number
      name:
        type: string
      tags:
        items:
          $ref: '#/definitions/repositories.CategoryTagTotal'
        type: array
    type: object
  repositories.DayExpense:
    properties:
      day:
        type: string
      day_name:
        type: string
      tags:
        items:
          $ref: '#/definitions/repositories.ExpenseTag'
        type: array
      total:
        type: number
    type: object
  repositories.ExpenseTag:
    properties:
      color:
        type: string
      name:
        type: string
      total:
        type: number
    type: object
  repositories.ExpensesMonthCurrentYear:
    properties:
      available_years:
        items:
          type: integer
        type: array
      months:
        items:
          $ref: '#/definitions/repositories.MonthCurrentYear'
        type: array
      total:
        type: number
      year:
        type: integer
    type: object
  repositories.MonthCurrentYear:
    properties:
      month:
        type: string
      total:
        type: number
    type: object
  repositories.MonthExpenses:
    properties:
      available_years:
        items:
          type: integer
        type: array
      month:
        type: string
      total_expenses:
        type: number
      weeks:
        items:
          $ref: '#/definitions/repositories.WeekExpenses'
        type: array
      year:
        type: integer
    type: object
  repositories.MonthOption:
    properties:
      label:
        type: string
      value:
        type: string
    type: object
  repositories.MonthlyCategoryExpense:
    properties:
      categories:
        items:
          $ref: '#/definitions/repositories.CategoryExpense'
        type: array
      month:
        type: string
      total:
        type: number
      year:
        type: integer
    type: object
  repositories.MonthlyTagExpense:
    properties:
      month:
        type: string
      tags:
        items:
          $ref: '#/definitions/repositories.TagExpense'
        type: array
      total:
        type: number
      year:
        type: integer
    type: object
  repositories.TagExpense:
    properties:
      tag_color:
        type: string
      tag_name:
        type: string
      total:
        type: number
    type: object
  repositories.WeekExpenses:
    properties:
      days:
        items:
          $ref: '#/definitions/repositories.DayExpense'
        type: array
      week:
        type: integer
    type: object
...
  util.ProblemDetails:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: guilherme.o.a.ufal@gmail.com
    name: Guilherme Amorim
    url: http://www.guilhermedeoliveiraamorim.com
  description: This is an API for managing expenses.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /expenses/available-months-years:
    get:
      description: Retrieves the list of months and years for which expense data is
        available for a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetAvailableMonthsYearsOutputDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get available months and years
      tags:
      - Utility
  /expenses/categories:
    get:
      description: Retrieves the expenses of a user categorized by category for a
        specified date range
      parameters:
      - description: Start date (DDMMYYYY)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (DDMMYYYY)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetExpensesByCategoryPeriodOutputDto'
        "400":
          description: Bad Request - Missing start date or end date
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get expenses by category for a period
      tags:
      - Presenters
  /expenses/categories/monthly:
    get:
      description: Retrieves the monthly expenses of a user categorized by category
        for a specific year
      parameters:
      - description: Year (YYYY)
        in: query
        name: year
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetMonthlyExpensesByCategoryYearOutputDto'
        "400":
          description: Bad Request - Missing or invalid year
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get monthly expenses by category for a specific year
      tags:
      - Presenters
  /expenses/day/day/period:
    get:
      description: A date range is passed to return a set of expenses in that range
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetDayToDayExpensesPeriodOutputDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get expenses in a time interval
      tags:
      - Presenters
  /expenses/monthly/total:
    get:
      description: Retrieves the total expenses of a user for the current month
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetTotalExpensesForCurrentMonthOutputDto'
        "400":
          description: Bad Request - Invalid parameters
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get total expenses for the current month
      tags:
      - Presenters
  /expenses/monthly/year:
    get:
      description: Retrieves expenses of a user for a specific month and year
      parameters:
      - description: Year of the expenses
        format: year
        in: query
        name: year
        required: true
        type: string
      - description: Month of the expenses
        format: month
        in: query
        name: month
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetExpensesByMonthYearOutputDto'
        "400":
          description: Bad Request - Missing or invalid parameters
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get expenses by month and year
      tags:
      - Presenters
  /expenses/tags/monthly:
    get:
      description: Retrieves the monthly expenses of a user categorized by tags for
        a specific year
      parameters:
      - description: Year (YYYY)
        in: query
        name: year
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetMonthlyExpensesByTagYearOutputDto'
        "400":
          description: Bad Request - Missing or invalid year
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get monthly expenses by tag for a specific year
      tags:
      - Presenters
  /expenses/tags/monthly/total:
    get:
      description: Retrieves the total expenses by category tags for a given month
        and year for a user
      parameters:
      - description: Year for which to retrieve expenses by category tags
        in: query
        name: year
        required: true
        type: string
      - description: Month for which to retrieve expenses by category tags
        in: query
        name: month
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetCategoryTagsTotalsByMonthYearOutputDto'
        "400":
          description: Bad Request - Missing or invalid year/month
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get totals of expenses by category tags for a specific month and year
      tags:
      - Presenters
  /expenses/total:
    get:
      description: Retrieves the total expenses of a user for a specified date range
      parameters:
      - description: Start date (DDMMYYYY)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (DDMMYYYY)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetTotalExpensesForPeriodOutputDto'
        "400":
          description: Bad Request - Missing start date or end date
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get total expenses for a period
      tags:
      - Presenters
  /expenses/total/monthly/year:
    get:
      description: Retrieves the total expenses for each month of the specified year
        for a user
      parameters:
      - description: Year for which to retrieve monthly expenses
        in: query
        name: year
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetTotalExpensesMonthCurrentYearOutputDto'
        "400":
          description: Bad Request - Missing or invalid year
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get total expenses for each month in the specified year
      tags:
      - Presenters
  /expenses/weekly/total:
    get:
      description: Retrieves the total expenses of a user for the current week
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.GetTotalExpensesForCurrentWeekOutputDto'
        "400":
          description: Bad Request - Invalid parameters
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ProblemDetails'
      security:
      - BearerAuth: []
      summary: Get total expenses for the current week
      tags:
      - Presenters
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"